# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/autores": {
            "get": {
                "description": "Obtém uma lista de todos os autores",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autores"
                ],
                "summary": "Retrieve all authors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Autores"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Adiciona um novo autor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autores"
                ],
                "summary": "Add a new author",
                "parameters": [
                    {
                        "description": "Author object",
                        "name": "autor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Autores"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Autores"
                        }
                    }
                }
            }
        },
        "/autores/{id}": {
            "get": {
                "description": "Obtém um autor de acordo com o ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autores"
                ],
                "summary": "Retrieve author by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Author ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Autores"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove um autor de acordo com o ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autores"
                ],
                "summary": "Delete author by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Author ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Autores"
                        }
                    }
                }
            }
        },
        "/categorias": {
            "get": {
                "description": "Obtém uma lista de todas as categorias",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categorias"
                ],
                "summary": "Retrieve all categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Categorias"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Adiciona uma nova categoria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categorias"
                ],
                "summary": "Add a new category",
                "parameters": [
                    {
                        "description": "Category object",
                        "name": "categoria",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Categorias"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Categorias"
                        }
                    }
                }
            }
        },
        "/categorias/{id}": {
            "get": {
                "description": "Obtém uma categoria de acordo com o ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categorias"
                ],
                "summary": "Retrieve category by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Categorias"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deleta uma categoria de acordo com o ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categorias"
                ],
                "summary": "Delete category by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/livros": {
            "get": {
                "description": "Obtém uma lista de todos os livros",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Livros"
                ],
                "summary": "Retrieve all books",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Livros"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Adiciona um novo livro",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Livros"
                ],
                "summary": "Add a new book",
                "parameters": [
                    {
                        "description": "Book object",
                        "name": "livro",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Livros"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Livros"
                        }
                    }
                }
            }
        },
        "/livros/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Livros"
                ],
                "summary": "Retrieve book by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Livros"
                        }
                    }
                }
            }
        },
        "/usuarios": {
            "get": {
                "description": "Obtém uma lista de todos os usuários",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Retrieve all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Usuarios"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Adiciona umo novo usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Add a new user",
                "parameters": [
                    {
                        "description": "User object",
                        "name": "usuario",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Usuarios"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Usuarios"
                        }
                    }
                }
            }
        },
        "/usuarios/{id}": {
            "get": {
                "description": "Obtém um usuário de acordo com o ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Retrieve user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Usuarios"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Autores": {
            "type": "object",
            "properties": {
                "Data_nascimento": {
                    "type": "string"
                },
                "ID_autor": {
                    "type": "integer"
                },
                "Nacionalidade": {
                    "type": "string"
                },
                "Nome": {
                    "type": "string"
                }
            }
        },
        "models.Categorias": {
            "type": "object",
            "properties": {
                "Descricao": {
                    "type": "string"
                },
                "ID_categoria": {
                    "type": "integer"
                }
            }
        },
        "models.Livros": {
            "type": "object",
            "properties": {
                "Ano_publicacao": {
                    "type": "integer"
                },
                "ID_autor": {
                    "type": "integer"
                },
                "ID_categoria": {
                    "type": "integer"
                },
                "ID_livro": {
                    "type": "integer"
                },
                "Isbn": {
                    "type": "string"
                },
                "Titulo": {
                    "type": "string"
                }
            }
        },
        "models.Usuarios": {
            "type": "object",
            "properties": {
                "Data_registro": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "ID_usuario": {
                    "type": "integer"
                },
                "Nome": {
                    "type": "string"
                },
                "Telefone": {
                    "type": "string"
                }
            }
        }
    }
}